refactor(a1): start core/helpers.py (pure helpers) + move cadences route (BR-CODE-1 slice 12)
Begins the shared-helper extraction. Moves the pure utility helpers
(_delta -> core.helpers.delta, _listings_cadence_seconds ->
listings_cadence_seconds) into core/helpers.py; app.py imports them aliased so
its inline usages (overview/section-metrics _delta) are unchanged. With the
cadence helper in core, moves /api/broker/event/{id}/cadences into
routers/broker.py (+test). 13 broker tests green; app.py -68 LOC (8211->8143).
