+++ core/helpers.py
@@ -1,0 +1,46 @@
+"""Pure utility helpers shared by app.py + routers/* (BR-CODE-1 helper pass).
+
+These have NO db/client/config dependencies — pure functions — so both app.py
+and the router modules import them from here (one-directional: app.py/routers ->
+core, never reverse). Moved verbatim from app.py.
+"""
+from __future__ import annotations
+
+from datetime import datetime
+
+
+def listings_cadence_seconds(occurs_at_local: str | None) -> int:
+    """Mirror collect-listings cron windows: closer events poll faster."""
+    if not occurs_at_local:
+        return 60 * 60 * 24
+    try:
+        # occurs_at_local is TEXT not TIMESTAMPTZ — known P1. Slice to date.
+        d = datetime.fromisoformat(occurs_at_local[:10])
+        days_out = (d - datetime.now()).days
+    except Exception:
+        return 60 * 60
+    if days_out <= 1:
+        return 60 * 20      # 20 min
+    if days_out <= 7:
+        return 60 * 60      # 60 min
+    if days_out <= 30:
+        return 60 * 60 * 4  # 4h
+    if days_out <= 60:
+        return 60 * 60 * 12  # 12h
+    return 60 * 60 * 24     # 24h
+
+
+def delta(curr, prev):
+    """Compute delta + percent for a numeric metric. Returns dict or None."""
+    if curr is None or prev is None:
+        return None
+    try:
+        c = float(curr); p = float(prev)
+    except (TypeError, ValueError):
+        return None
+    if c == p:
+        return {"abs": 0, "pct": 0, "dir": "flat"}
+    diff = c - p
+    pct = (diff / p * 100) if p != 0 else None
+    return {"abs": round(diff, 2), "pct": round(pct, 2) if pct is not None else None,
+            "dir": "up" if diff > 0 else "down"}
